Separate LinearEmbeddingEvaluation from EmbeddingEvaluatorModel.py
