+++ ms2deepscore/models/__init__.py
@@ -1,8 +1,8 @@
 from .load_model import load_model, load_embedding_evaluator, load_linear_model
 from .loss_functions import bin_dependent_losses
 from .EmbeddingEvaluatorModel import (EmbeddingEvaluationModel,
-                                      LinearModel,
                                       train_evaluator)
+from .LinearEmbeddingEvaluation import LinearModel
 from .SiameseSpectralModel import (SiameseSpectralModel,
                                    compute_embedding_array, train)
 
@@ -11,7 +11,6 @@ __all__ = [
     "bin_dependent_losses",
     "compute_embedding_array",
     "EmbeddingEvaluationModel",
-    "LinearModel",
     "load_embedding_evaluator",
     "load_linear_model",
     "load_model",
