+++ ms2deepscore/models/load_model.py
@@ -4,7 +4,8 @@ from typing import Union
 import numpy as np
 import torch
 from ms2deepscore.__version__ import __version__
-from ms2deepscore.models.EmbeddingEvaluatorModel import EmbeddingEvaluationModel, LinearModel
+from ms2deepscore.models.EmbeddingEvaluatorModel import EmbeddingEvaluationModel
+from ms2deepscore.models import LinearModel
 from ms2deepscore.models.SiameseSpectralModel import SiameseSpectralModel
 from ms2deepscore.SettingsMS2Deepscore import SettingsMS2Deepscore
